Fetch all collections from an rstac::collections() query

By default, NASA STAC catalogue collections queries only return 10 collections at a time. This function will page through the collections and return them all.

Usage

collections_fetch(collections, ...)

Arguments

collections

an object of class doc_collections as returned by calling rstac::get_request() on an rstac::collections() query

...

Optional arguments passed on to httr::GET()

Value

A doc_collections object with all of the collections from the catalogue specified in the rstac::collections() query.

Examples

if (FALSE) { # interactive()

 rstac::stac("https://cmr.earthdata.nasa.gov/stac/LPCLOUD") |> 
   rstac::collections() |> 
   rstac::get_request() |> 
   collections_fetch()
}